+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [0.0.10] 2022-03-16
+
+### Added
+
+- Support for custom plugin parameter types
+- Enumeration parameter type
+
 ## [0.0.9] 2022-03-09
 
 ### Fixed